Unit Test

See merge request HCM24_CPL_JAVA_04_TranDangQuocDat/cursus!9

+++ src/main/java/com/fpt/cursus/repository/BlackListIpRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface BlackListIpRepo extends JpaRepository<BlackListIp, Long> {
+public interface BlackListIpRepo extends JpaRepository<BlackListIp, Integer> {
     Optional<BlackListIp> findByIpAddress(String ipAddress);
 }
